feat : accessToken 만료시 refreshToken으로 accessToken 갱신시키는 기능 추가

+++ src/main/java/com/example/to_do_list/domain/Attend.java
@@ -19,14 +19,14 @@ public class Attend {
 
     private LocalDate date;
     private Long userId;
-    private int percentage;
+    private double percentage;
 
     @ManyToOne
     @JoinColumn(name = "team_id")
     private Team team;
 
     @Builder
-    public Attend(LocalDate date, Long userId, int percentage, Team team) {
+    public Attend(LocalDate date, Long userId, double percentage, Team team) {
         this.date = date;
         this.userId = userId;
         this.percentage = percentage;
